Updating docs and making ONNXAttribute a top level class.
