+++ Util/ONNXExport/src/main/java/org/tribuo/onnx/ONNXOperators.java
@@ -16,10 +16,8 @@
 
 package org.tribuo.onnx;
 
-import com.google.protobuf.ByteString;
 import onnx.OnnxMl;
 
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -70,7 +68,7 @@ public enum ONNXOperators {
         this.numOutputs = numOutputs;
         Map<String,ONNXAttribute> attributeMap = new HashMap<>();
         for (ONNXAttribute a : attributes) {
-            attributeMap.put(a.name,a);
+            attributeMap.put(a.getName(),a);
         }
         if (attributes.size() != attributeMap.size()) {
             throw new IllegalArgumentException("Duplicate attribute in enum declaration - " + attributes);
@@ -115,118 +113,6 @@ public enum ONNXOperators {
             nodeBuilder.addAttribute(attr.build(e.getValue()));
         }
         return nodeBuilder.build();
-    }
-
-    /**
-     * The spec for an attribute, used to produce the attribute proto at construction time.
-     */
-    public static final class ONNXAttribute {
-        private final String name;
-        private final OnnxMl.AttributeProto.AttributeType type;
-
-        /**
-         * Constructs an attribute placeholder of the appropriate name and type.
-         * @param name The name of the attribute.
-         * @param type The type of the attribute.
-         */
-        public ONNXAttribute(String name, OnnxMl.AttributeProto.AttributeType type) {
-            this.name = name;
-            this.type = type;
-        }
-
-        /**
-         * Gets the attribute's name.
-         * @return The attribute's name.
-         */
-        public String getName() {
-            return name;
-        }
-
-        /**
-         * Builds the attribute proto using the supplied value.
-         * <p>
-         * Throws {@link IllegalArgumentException} if the value type does not match the expected type, and
-         * throws {@link UnsupportedOperationException} if the value type is not supported. Currently supported
-         * types are primitives, strings and arrays of primitives and strings.
-         * @param value The value
-         * @return The AttributeProto.
-         */
-        public OnnxMl.AttributeProto build(Object value) {
-            OnnxMl.AttributeProto.Builder builder = OnnxMl.AttributeProto.newBuilder();
-
-            builder.setName(name);
-            builder.setType(type);
-            switch (type) {
-                case FLOAT:
-                    if (value instanceof Float) {
-                        builder.setF((Float)value);
-                    } else {
-                        throw new IllegalArgumentException("Expected Float, found " + value.getClass() + " with value " + value);
-                    }
-                    break;
-                case INT:
-                    if (value instanceof Integer) {
-                        builder.setI((Integer)value);
-                    } else {
-                        throw new IllegalArgumentException("Expected Integer, found " + value.getClass() + " with value " + value);
-                    }
-                    break;
-                case STRING:
-                    if (value instanceof String) {
-                        builder.setS(ByteString.copyFrom((String)value, StandardCharsets.UTF_8));
-                    } else {
-                        throw new IllegalArgumentException("Expected String, found " + value.getClass() + " with value " + value);
-                    }
-                    break;
-                case FLOATS:
-                    if (value instanceof float[]) {
-                        float[] arr = (float[]) value;
-                        for (int i = 0; i < arr.length; i++) {
-                            builder.addFloats(arr[i]);
-                        }
-                    } else {
-                        throw new IllegalArgumentException("Expected float[], found " + value.getClass() + " with value " + value);
-                    }
-                    break;
-                case INTS:
-                    if (value instanceof int[]) {
-                        int[] arr = (int[]) value;
-                        for (int i = 0; i < arr.length; i++) {
-                            builder.addInts(arr[i]);
-                        }
-                    } else {
-                        throw new IllegalArgumentException("Expected int[], found " + value.getClass() + " with value " + value);
-                    }
-                    break;
-                case STRINGS:
-                    if (value instanceof String[]) {
-                        String[] arr = (String[]) value;
-                        for (int i = 0; i < arr.length; i++) {
-                            builder.addStrings(ByteString.copyFrom(arr[i], StandardCharsets.UTF_8));
-                        }
-                    } else {
-                        throw new IllegalArgumentException("Expected String[], found " + value.getClass() + " with value " + value);
-                    }
-                    break;
-                case UNDEFINED:
-                case TENSOR:
-                case GRAPH:
-                case SPARSE_TENSOR:
-                case TYPE_PROTO:
-                case TENSORS:
-                case GRAPHS:
-                case SPARSE_TENSORS:
-                case TYPE_PROTOS:
-                    throw new UnsupportedOperationException("Type: " + type + " is not supported.");
-            }
-
-            return builder.build();
-        }
-
-        @Override
-        public String toString() {
-            return "ONNXAttribute(name='"+name+"',type="+type+")";
-        }
     }
 
 }
